requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-environ==0.11.2
django-model-utils==5.0.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
dj-rest-auth==6.0.0
drf-spectacular==0.27.1
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
requests==2.32.3
pillow==11.2.1
dateutils==0.6.12
